requirements.txt
appdirs==1.4.0
click==6.7
decorator==4.0.11
Flask==0.12
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
pbr==3.0.0
pg8000==1.10.6
psycopg2==2.6.2
pyparsing==2.1.10
six==1.10.0
SQLAlchemy==1.1.9
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.11.15
WTForms==2.1
